+++ moncic/ci.py
@@ -137,8 +137,8 @@ class CI(MoncicCommand):
                             help="branch to be used. Default: let 'git clone' choose")
         parser.add_argument("-s", "--system", action="store",
                             help="name or path of the system used to build")
-        parser.add_argument("-b", "--build-style", action="store", default="travis",
-                            help="name of the procedure used to run the CI. Default: 'travis'")
+        parser.add_argument("-b", "--build-style", action="store",
+                            help="name of the procedure used to run the CI. Default: autodetect")
         parser.add_argument("-a", "--artifacts", metavar="dir", action="store",
                             help="directory where build artifacts will be stored")
         parser.add_argument("--shell", action="store_true",